Move interfaces to its own files
