+++ src/EventSourced.Persistence.EntityFramework/Mappers/AggregateBasedProjectionEntityMapper.cs
@@ -6,12 +6,6 @@ using Newtonsoft.Json;
 
 namespace EventSourced.Persistence.EntityFramework.Mappers
 {
-    public interface IAggregateBasedProjectionEntityMapper
-    {
-        AggregateBasedProjectionEntity MapToEntity(Guid aggregateRootId, object projection);
-        object MapToProjection(AggregateBasedProjectionEntity entity);
-    }
-
     public class AggregateBasedProjectionEntityMapper : IAggregateBasedProjectionEntityMapper
     {
         private readonly ITypeSerializer _typeSerializer;
