+++ src/EventSourced.Persistence.EntityFramework/Mappers/IAggregateBasedProjectionEntityMapper.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using EventSourced.Persistence.EntityFramework.Entities;
+
+namespace EventSourced.Persistence.EntityFramework.Mappers
+{
+    public interface IAggregateBasedProjectionEntityMapper
+    {
+        AggregateBasedProjectionEntity MapToEntity(Guid aggregateRootId, object projection);
+        object MapToProjection(AggregateBasedProjectionEntity entity);
+    }
+}
