+++ src/EventSourced.Persistence.EntityFramework/Mappers/ITypeBaseProjectionEntityMapper.cs
@@ -1,0 +1,10 @@
+﻿using EventSourced.Persistence.EntityFramework.Entities;
+
+namespace EventSourced.Persistence.EntityFramework.Mappers
+{
+    public interface ITypeBaseProjectionEntityMapper
+    {
+        TypeBasedProjectionEntity MapToEntity(object projection);
+        object MapToProjection(TypeBasedProjectionEntity entity);
+    }
+}
