+++ src/EventSourced.Persistence.EntityFramework/Mappers/TypeBaseProjectionEntityMapper.cs
@@ -4,12 +4,6 @@ using Newtonsoft.Json;
 
 namespace EventSourced.Persistence.EntityFramework.Mappers
 {
-    public interface ITypeBaseProjectionEntityMapper
-    {
-        TypeBasedProjectionEntity MapToEntity(object projection);
-        object MapToProjection(TypeBasedProjectionEntity entity);
-    }
-
     public class TypeBaseProjectionEntityMapper : ITypeBaseProjectionEntityMapper
     {
         private readonly ITypeSerializer _typeSerializer;
